+++ R/lin-alg.R
@@ -208,6 +208,32 @@ reciprocal_matrix <- function(x, numerator = 1){
 }
 
 
+#' Matrix power
+#'
+#' @param x A `caracas_symbol`, a matrix.
+#' @param pow 
+#' 
+#' @examples
+#' if (has_sympy()) {
+#'   M <- matrix_(c("1", "a", "a", 1), 2, 2)
+#'   M
+#'   mat_pow(M, 1/2)
+#' }
+#' 
+#' @concept linalg
+#' @export
+mat_pow <- function(x, pow = "1") {
+  ensure_sympy()
+  stopifnot_symbol(x)
+  stopifnot_matrix(x)
+  
+  pow_str <- deparse1(substitute(pow)) # to avoid 1/2 gets converted to 0.5
+
+  x_pow <- x$pyobj$pow(pow_str)
+  
+  return(construct_symbol_from_pyobj(x_pow))
+}
+
 
 #' Matrix diagonal
 #'